VBID-4331 ability to specify class equality
use SomeClass.__honk_clsid as an alternate way to specify class
equality.

This came up in instances where a project using honk-di depended on
another project using honk-di, where both share a dependency that has an
abstract class, and in both projects that abstract class is bound to an
implementation

this allows the setting of __honk_clsid to say "alright, these are
actually the same thing, this is fine".  this might be a terrible idea

* whitespace cleanup
* move gulpfile.js -> gulpfile.coffee